remove CCA_IN_AD config flag
