+++ cmd/agent/common/autodiscovery.go
@@ -18,7 +18,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/scheduler"
 	"github.com/DataDog/datadog-agent/pkg/config"
 	confad "github.com/DataDog/datadog-agent/pkg/config/autodiscovery"
-	"github.com/DataDog/datadog-agent/pkg/util"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
@@ -72,28 +71,16 @@ func setupAutoDiscovery(confSearchPaths []string, metaScheduler *scheduler.MetaS
 			}
 		}
 
-		ccaInAD := util.CcaInAD()
-		if ccaInAD {
-			var enableContainerProvider bool
-			for _, p := range incompatibleProviders {
-				if _, found := uniqueConfigProviders[p]; found {
-					enableContainerProvider = true
-					delete(uniqueConfigProviders, p)
-				}
+		var enableContainerProvider bool
+		for _, p := range incompatibleProviders {
+			if _, found := uniqueConfigProviders[p]; found {
+				enableContainerProvider = true
+				delete(uniqueConfigProviders, p)
 			}
+		}
 
-			if enableContainerProvider {
-				uniqueConfigProviders[names.KubeContainer] = config.ConfigurationProviders{Name: names.KubeContainer}
-			}
-		} else {
-			// The "docker" config provider was replaced with the "container" one
-			// that supports Docker, but also other runtimes. We need this
-			// conversion to avoid breaking configs that included "docker".
-			if options, found := uniqueConfigProviders["docker"]; found {
-				delete(uniqueConfigProviders, "docker")
-				options.Name = names.Container
-				uniqueConfigProviders["container"] = options
-			}
+		if enableContainerProvider {
+			uniqueConfigProviders[names.KubeContainer] = config.ConfigurationProviders{Name: names.KubeContainer}
 		}
 
 		for _, provider := range extraConfigProviders {
