+++ pkg/autodiscovery/autoconfig.go
@@ -21,8 +21,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/status/health"
 	"github.com/DataDog/datadog-agent/pkg/tagger"
-	"github.com/DataDog/datadog-agent/pkg/util"
-	"github.com/DataDog/datadog-agent/pkg/util/containers"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/util/retry"
 )
@@ -75,12 +73,7 @@ func NewAutoConfig(scheduler *scheduler.MetaScheduler) *AutoConfig {
 
 // NewAutoConfigNoStart creates an AutoConfig instance.
 func NewAutoConfigNoStart(scheduler *scheduler.MetaScheduler) *AutoConfig {
-	var cfgMgr configManager
-	if util.CcaInAD() {
-		cfgMgr = newReconcilingConfigManager()
-	} else {
-		cfgMgr = newSimpleConfigManager()
-	}
+	cfgMgr := newReconcilingConfigManager()
 	ac := &AutoConfig{
 		configPollers:      make([]*configPoller, 0, 9),
 		listenerCandidates: make(map[string]*listenerCandidate),
@@ -423,18 +416,6 @@ func (ac *AutoConfig) processNewService(ctx context.Context, svc listeners.Servi
 	}
 
 	changes := ac.cfgMgr.processNewService(ADIdentifiers, svc)
-
-	if !util.CcaInAD() {
-		// schedule a "service config" for logs-agent's benefit
-		changes.ScheduleConfig(integration.Config{
-			LogsConfig:      integration.Data{},
-			ServiceID:       svc.GetServiceID(),
-			TaggerEntity:    svc.GetTaggerEntity(),
-			MetricsExcluded: svc.HasFilter(containers.MetricsFilter),
-			LogsExcluded:    svc.HasFilter(containers.LogsFilter),
-		})
-	}
-
 	ac.applyChanges(changes)
 }
 
@@ -443,18 +424,6 @@ func (ac *AutoConfig) processDelService(ctx context.Context, svc listeners.Servi
 	ac.store.removeServiceForEntity(svc.GetServiceID())
 	changes := ac.cfgMgr.processDelService(ctx, svc)
 	ac.store.removeTagsHashForService(svc.GetTaggerEntity())
-
-	if !util.CcaInAD() {
-		// unschedule the "service config"
-		changes.UnscheduleConfig(integration.Config{
-			LogsConfig:      integration.Data{},
-			ServiceID:       svc.GetServiceID(),
-			TaggerEntity:    svc.GetTaggerEntity(),
-			MetricsExcluded: svc.HasFilter(containers.MetricsFilter),
-			LogsExcluded:    svc.HasFilter(containers.LogsFilter),
-		})
-	}
-
 	ac.applyChanges(changes)
 }
