+++ pkg/autodiscovery/configmgr_test.go
@@ -414,10 +414,6 @@ func (suite *ConfigManagerSuite) TestFuzz() {
 	})
 }
 
-func TestSimpleConfigManagement(t *testing.T) {
-	suite.Run(t, &ConfigManagerSuite{factory: newSimpleConfigManager})
-}
-
 type ReconcilingConfigManagerSuite struct {
 	ConfigManagerSuite // include all ConfigManager tests, and more..
 }
