+++ pkg/autodiscovery/providers/kubelet.go
@@ -18,7 +18,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/status/health"
-	"github.com/DataDog/datadog-agent/pkg/util"
 	"github.com/DataDog/datadog-agent/pkg/util/containers"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/workloadmeta"
@@ -180,9 +179,7 @@ func (k *KubeletConfigProvider) generateConfigs() ([]integration.Config, error) 
 				}
 			}
 
-			if util.CcaInAD() {
-				c = utils.AddContainerCollectAllConfigs(c, containerEntity)
-			}
+			c = utils.AddContainerCollectAllConfigs(c, containerEntity)
 
 			containerIdentifiers[adIdentifier] = struct{}{}
 			containerNames[podContainer.Name] = struct{}{}
