+++ pkg/logs/logs.go
@@ -23,10 +23,8 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/logs/config"
 	"github.com/DataDog/datadog-agent/pkg/logs/diagnostic"
 	adScheduler "github.com/DataDog/datadog-agent/pkg/logs/schedulers/ad"
-	ccaScheduler "github.com/DataDog/datadog-agent/pkg/logs/schedulers/cca"
 	"github.com/DataDog/datadog-agent/pkg/logs/service"
 	"github.com/DataDog/datadog-agent/pkg/logs/status"
-	ddUtil "github.com/DataDog/datadog-agent/pkg/util"
 )
 
 const (
@@ -135,9 +133,6 @@ func start(ac *autodiscovery.AutoConfig, serverless bool) (*Agent, error) {
 			panic("AutoConfig must be initialized before logs-agent")
 		}
 		agent.AddScheduler(adScheduler.New(ac))
-		if !ddUtil.CcaInAD() {
-			agent.AddScheduler(ccaScheduler.New(ac))
-		}
 	}
 
 	return agent, nil
